backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
requests>=2.26.0
pandas>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
